fix and add unit test

+++ src/main/kotlin/com/example/challengeCupon/challengeCupon/adapter/model/CouponCalculateResponse.kt
@@ -1,17 +1,18 @@
 package com.example.challengeCupon.challengeCupon.adapter.model
 
 import com.example.challengeCupon.challengeCupon.application.usecase.model.CouponCalculate
+import java.math.BigDecimal
 
 class CouponCalculateResponse(
     private var itemsIds: List<String>,
-    private var total: Float
+    private var total: BigDecimal
 ) {
 
     fun getItemsIds(): List<String>{
         return itemsIds
     }
 
-    fun getTotal(): Float {
+    fun getTotal(): BigDecimal {
         return total
     }
