+++ src/main/kotlin/com/example/challengeCupon/challengeCupon/application/port/in/CalculateCouponCommand.kt
@@ -1,6 +1,7 @@
 package com.example.challengeCupon.challengeCupon.application.port.`in`
 
 import com.example.challengeCupon.challengeCupon.application.usecase.model.CouponCalculate
+import java.math.BigDecimal
 
 interface CalculateCouponCommand {
 
@@ -8,14 +9,14 @@ interface CalculateCouponCommand {
 
     class Command(
         private var itemsIds: List<String>,
-        private var amount: Float
+        private var amount: BigDecimal
     ) {
 
         fun getItemsIds(): List<String> {
             return itemsIds
         }
 
-        fun getAmount(): Float {
+        fun getAmount(): BigDecimal {
             return amount
         }
     }
