+++ src/main/kotlin/com/example/challengeCupon/challengeCupon/application/usecase/model/Coupon.kt
@@ -1,14 +1,16 @@
 package com.example.challengeCupon.challengeCupon.application.usecase.model
 
+import java.math.BigDecimal
+
 class Coupon (
     private var itemsIds: List<String>,
-    private var amount: Float
+    private var amount: BigDecimal
 ) {
     fun getItemsIds(): List<String>{
         return itemsIds
     }
 
-    fun getAmount(): Float {
+    fun getAmount(): BigDecimal {
         return amount
     }
 }
