+++ src/main/kotlin/com/example/challengeCupon/challengeCupon/application/usecase/model/CouponCalculate.kt
@@ -1,14 +1,16 @@
 package com.example.challengeCupon.challengeCupon.application.usecase.model
 
+import java.math.BigDecimal
+
 class CouponCalculate(
     private var itemsIds: List<String>,
-    private var total: Float
+    private var total: BigDecimal
 ) {
     fun getItemsIds(): List<String>{
         return itemsIds
     }
 
-    fun getTotal(): Float {
+    fun getTotal(): BigDecimal {
         return total
     }
 }
