+++ src/main/kotlin/com/example/challengeCupon/challengeCupon/application/usecase/model/Item.kt
@@ -1,15 +1,17 @@
 package com.example.challengeCupon.challengeCupon.application.usecase.model
 
+import java.math.BigDecimal
+
 data class Item(
     private var id: String,
-    private var price: Float
+    private var price: BigDecimal
 ) {
 
     fun getId(): String {
         return id
     }
 
-    fun getPrice(): Float {
+    fun getPrice(): BigDecimal {
         return price
     }
